tesla-energy-project/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.5
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
python-multipart>=0.0.6
aiohttp>=3.9.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

